// src/common/emums/urls.enum.ts
export enum UrlsEnum {
  clientAppPlaystoreUrl = "https://play.google.com",
  riderAppPlaystoreUrl = "https://play.google.com",
  localHostBaseUrl = "http://localhost:3031/",
  prodServerBaseUrl = "https://prod.co.ke/images/",
  testServerBaseUrl = "https://test.co.ke/",
  productionServerBaseUrl = "https://prodserver.co.ke/",
  NGROK_URL = "https://575ae2e5.ngrok.io",
}