// src/graphql.schema.ts
export class CompanyStaffStats {
    total?: number;
    totalMeterReader?: number;
    totalItManager?: number;
    totalAccountant?: number;
    totalManager?: number;
    totalStaff?: number;
    totalActive?: number;
    totalInactive?: number;
    totalSecretary?: number;
    totalSoftDeleted?: number;
}