// src/modules/customer/dto/reset-password.dto.ts
export class ResetPassword {
  customerId?: number;
  companyStaffId?: number;
  companyId?: number;
  phoneNo?: string;
  email?: string;
}

export class OtpResetPassword {
  otpCode?: string;
  phoneNo?: string;
  password?: string;
}

export class ChangePassword {
  companyStaffId?: number;
  driverId?: number;
  newPassword?: string;
  oldPassword?: string;
  companyId: number;
  email?: string;
}